# API Interface/notification-center-swagger (20-01-2025).yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Notifacation Center API",
    "description": "API documentation for Notifacation Center service",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8091",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/notifications": {
      "get": {
        "tags": [
          "notification-controller"
        ],
        "summary": "Retrieve all Notifications",
        "operationId": "getAllNotifications",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortAttribute",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "timestamp"
            }
          },
          {
            "name": "isAscending",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications retrieved successfully!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponsePaginatedResultsDtoNotificationDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "401": {
            "description": "Authentication process failed!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "403": {
            "description": "Invalid authorization parameters. Check JWT or CSRF Token",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "404": {
            "description": "Invalid sort attributes",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponsePaginatedResultsDtoNotificationDto"
                }
              }
            }
          },
          "417": {
            "description": "Expectation Failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      }
    },
    "/api/notifications/{notificationId}": {
      "get": {
        "tags": [
          "notification-controller"
        ],
        "summary": "Retrieve Notification By Id",
        "operationId": "getNotificationById",
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications retrieved successfully!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseNotificationDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "401": {
            "description": "Authentication process failed!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "403": {
            "description": "Invalid authorization parameters. Check JWT or CSRF Token",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "417": {
            "description": "Notification with id [ID] not found in DB",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      }
    },
    "/api/notifications/user/{userId}": {
      "get": {
        "tags": [
          "notification-controller"
        ],
        "summary": "Retrieve all notification per UserID",
        "operationId": "getAllNotificationPerUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortAttribute",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "timestamp"
            }
          },
          {
            "name": "isAscending",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications retrieved successfully!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponsePaginatedResultsDtoNotificationDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "401": {
            "description": "Authentication process failed!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "403": {
            "description": "Invalid authorization parameters. Check JWT or CSRF Token",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "417": {
            "description": "Expectation Failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      }
    },
    "/api/notifications/user/{userId}/unread": {
      "get": {
        "tags": [
          "notification-controller"
        ],
        "summary": "Retrieve all unread notifications for a specific user",
        "operationId": "getAllUnreadNotificationPerUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unread notifications retrieved successfully!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseListNotificationDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "401": {
            "description": "Authentication process failed!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "403": {
            "description": "Invalid authorization parameters. Check JWT or CSRF Token",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "417": {
            "description": "Expectation Failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "BaseAppResponseString": {
        "type": "object",
        "properties": {
          "data": {
            "type": "string"
          },
          "errors": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BaseAppResponseMapStringString": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "errors": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "BaseAppResponsePaginatedResultsDtoNotificationDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PaginatedResultsDtoNotificationDto"
          },
          "errors": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "NotificationDto": {
        "type": "object",
        "properties": {
          "notificationId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "notificationType": {
            "type": "string",
            "enum": [
              "Event",
              "Assignment"
            ]
          },
          "relatedEvent": {
            "type": "string"
          },
          "relatedAssignment": {
            "type": "string"
          },
          "notificationStatus": {
            "type": "string",
            "enum": [
              "NOT_VIEWED",
              "VIEWED"
            ]
          },
          "sourceComponent": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "productionModule": {
            "type": "string"
          },
          "smartService": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          }
        }
      },
      "PaginatedResultsDtoNotificationDto": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationDto"
            }
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int32"
          },
          "lastPage": {
            "type": "boolean"
          }
        }
      },
      "BaseAppResponseNotificationDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/NotificationDto"
          },
          "errors": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BaseAppResponseListNotificationDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationDto"
            }
          },
          "errors": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}